Refactor: extract font management to FontsManager
- Created FontsManager: ObservableObject to handle user font registration and management.
- Migrated userFontInfos, loadUserFonts, importCustomFonts, and removeCustomFonts logic.
- Updated ModelData to delegate font management to FontsManager while keeping a bridge for userFontInfos.
- Decoupled ReaderOptionsView by directly observing FontsManager from the environment.
- Deleted the now-redundant FontsHelper.swift.
